refactor(playground): extract state logic into focused hooks
Extract compilation, Monaco sync, debounced compile, and editor actions
into dedicated hooks. Move compile utilities to a separate module.

- playground.tsx reduced from ~600 to ~360 lines
- No public API changes (backward-compatible)
- All existing tests pass
